requirements.txt
# AI Quote Reel Bot - Google AI Stack
# Requirements for automated reel generation

# Google AI SDK (Gemini 2.5)
google-genai>=1.0.0

# Google Cloud (Imagen 3 + TTS)
google-cloud-aiplatform>=1.38.0
google-cloud-texttospeech>=2.14.0

# Video Processing
moviepy>=1.0.3
Pillow>=10.0.0
numpy>=1.24.0

# Audio Processing
pydub>=0.25.1
mutagen>=1.47.0

# YouTube Upload
google-api-python-client>=2.100.0
google-auth>=2.23.0
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.1.1

# Utils
python-dotenv>=1.0.0
httpx>=0.25.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
loguru>=0.7.0
